# src/integration-test/resources/openapi.yml
openapi: 3.1.0
info:
  title: api-service-memories
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "3.12"
servers:
- url: http://localhost:61485
  description: Generated server url
...
paths:
...
  /{municipalityId}/objects:
    get:
      tags:
      - Object
      summary: Search objects
      description: "Search across all object types (Foto, Föremål, Film, Ljud, Text,\
        \ Publikation) and registers (Person, Juridisk person, Sjöman, Mantal) in\
        \ one call, with global sorting, pagination and counts per type, gender, originator\
        \ category and place. Pass objectType, gender, categoryId or topographyId\
        \ to narrow the search; each counter leaves out only its own selection, so\
        \ the chips stay selectable. Pass nodeId to list what an archive node holds."
      operationId: searchObjects
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: parameters
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/CombinedObjectParameters"
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedCombinedObjectResponse"
  /{municipalityId}/nodes:
    get:
      tags:
      - Node
      summary: Search archive and collection nodes
      description: "Search the archive tree (arkiv, serier, volymer, samlingar) by\
        \ free text, node type (by id or by name — nodeType=Arkiv is the list of arkiv\
        \ och samlingar), period, institution, category and place. Every node carries\
        \ what the archive records about it — institution, arkivbildare, place, signum\
        \ and so on — and the objects it holds are listed by /objects?nodeId=. Returns\
        \ paginated results."
      operationId: searchNodes
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: parameters
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NodeParameters"
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedNodeResponse"
  /{municipalityId}/nodes/{id}:
    get:
      tags:
      - Node
      summary: Get node by ID
      description: "Get a specific archive or collection node together with the path\
        \ from the root down to it, for a breadcrumb."
      operationId: getNode
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/nodes/{id}/children:
    get:
      tags:
      - Node
      summary: Get the children of a node
      description: "List the nodes directly below a node, in the archive's own order\
        \ unless another sort is requested. Takes the same filters as the search."
      operationId: getNodeChildren
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: parameters
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NodeParameters"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedNodeResponse"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Direction:
      type: string
      enum:
      - ASC
      - DESC
...
    CombinedObjectParameters:
      type: object
      description: "Combined object search parameters (across all object and register\
        \ types). All filters are optional and combined with AND, except that several\
        \ values of objectType are alternatives. Sort on one of: relevance, objectKey,\
        \ title, year, objectType or location. Defaults to relevance when a query\
        \ is given."
      properties:
        page:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - 1
          minimum: 1
        limit:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - 15
          minimum: 1
        sortBy:
          type: array
          items:
            type: string
            description: "Property to sort on. 'relevance' ranks the best match first\
              \ and is the default when a query is given; it is ignored without one.\
              \ 'location' sorts on the free-text location, which for persons and\
              \ seamen holds the birth parish. Sorting is ascending by default, which\
              \ for relevance means most relevant first."
            enum:
            - relevance
            - objectKey
            - title
            - year
            - objectType
            - location
            examples:
            - relevance
        sortDirection:
          $ref: "#/components/schemas/Direction"
          description: The sort order direction
          examples:
          - ASC
        query:
          type: string
          description: "Free text search (case-insensitive). Every word must occur\
            \ somewhere in the title and comment, and for the register types also\
            \ in names, parishes and other identifying fields, in any order. Results\
            \ are ranked with title and name matches above matches that only occurred\
            \ in a comment or a body text."
          examples:
          - Anton Nordin
        yearFrom:
          type: integer
          format: int32
          description: Year from (inclusive)
          examples:
          - 1900
        yearTo:
          type: integer
          format: int32
          description: Year to (inclusive)
          examples:
          - 1950
        location:
          type: string
          description: "Location (substring, case-insensitive; resolved place name\
            \ or free-text location)"
          examples:
          - Sundsvall
        topographyId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the place, as listed by /topographies — the exact\
              \ counterpart of location, for a place picked from that list rather\
              \ than typed, and counted by topographyCounts. Repeat the parameter,\
              \ or comma-separate the values, to select several — they are alternatives.\
              \ Only the object types and legal entities are placed in a topography;\
              \ the person registers hold a parish as free text, so this filter excludes\
              \ them."
            examples:
            - 1
        objectType:
          type: array
          items:
            type: string
            description: "Object type to include: Foto, Föremål, Film, Ljud, Text,\
              \ Publikation, Person, Juridisk person, Sjöman or Mantal. Repeat the\
              \ parameter, or comma-separate the values, to select several — they\
              \ are alternatives, so Foto and Ljud together means either. Omit to\
              \ include every type. The values are the ones each object reports under\
              \ objectType and typeCounts counts by, so a client can filter on exactly\
              \ what it counted."
            examples:
            - Foto
        gender:
          type: string
          description: "Gender — Man, Kvinna or Okänt, matched case-insensitively.\
            \ Only the person registers record one, and the filter speaks for those\
            \ types alone: it narrows them to the chosen gender and leaves every other\
            \ type untouched, so combining it with objectType=Foto returns the photos\
            \ and the matching persons rather than nothing. Okänt covers every register\
            \ row without a readable gender, so the three values account for all of\
            \ them."
          examples:
          - Man
        creator:
          type: string
          description: "Originator (upphovsman) name (substring, case-insensitive;\
            \ matches a person or a legal entity). Only object types carry an originator,\
            \ so this filter also excludes the register types."
          examples:
          - Nordin
        creatorPersonId:
          type: integer
          format: int32
          description: "ID of the originator, when it is a person"
          examples:
          - 1
        creatorLegalEntityId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the originator, when it is a legal entity. Repeat\
              \ the parameter, or comma-separate the values, to select several — they\
              \ are alternatives, so a category of legal entities can be filtered\
              \ in one call."
            examples:
            - 10
        categoryId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the originator's category (verksamhetskategori), as\
              \ listed by /categories and counted by categoryCounts. Repeat the parameter,\
              \ or comma-separate the values, to select several — they are alternatives.\
              \ A category is a property of the originator, so like creator this filter\
              \ keeps only the objects with an originator in it and excludes the register\
              \ types."
            examples:
            - 5
        nodeId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the archive node (arkiv, serie or volym) to list the\
              \ objects of, as returned by /nodes — what the node's own page shows\
              \ under it. Repeat the parameter, or comma-separate the values, to select\
              \ several — they are alternatives. An object sits in exactly one node,\
              \ so a node's descendants are not searched; list them with /nodes/{id}/children.\
              \ Only the object types are placed in the tree, so this filter excludes\
              \ the register types."
            examples:
            - 19000
    CategoryCount:
      type: object
      description: How many objects with an originator in one category the search
        matches
      properties:
        categoryId:
          type: integer
          format: int32
          description: "Category ID, which the categoryId filter accepts"
          examples:
          - 5
        name:
          type: string
          description: Category name
          examples:
          - Förening
        count:
          type: integer
          format: int64
          description: "Number of matching objects whose originator is a legal entity\
            \ in that category, across every page. The counters cover every matched\
            \ row with a categorised originator, so they sum to fewer than totalRecords\
            \ whenever the result also holds rows without one."
          examples:
          - 12
    CombinedObject:
      type: object
      description: Combined object search hit (any object or register type)
      properties:
        objectKey:
          type: string
          description: "Stable key across types ({type}-{id}; census records key on\
            \ the volume and the row number, e.g. mantal-1845-123)"
          examples:
          - foto-1001
        sourceId:
          type: integer
          format: int32
          description: Source id within its own type
          examples:
          - 1001
        objectType:
          type: string
          description: "Object type (Foto, Föremål, Film, Ljud, Text, Publikation,\
            \ Person, Juridisk person, Sjöman or Mantal)"
          examples:
          - Foto
        title:
          type: string
          description: "Title — the document title for objects, the composed name\
            \ for Person, Juridisk person and Sjöman"
          examples:
          - Stadsvy från Norra berget
        year:
          type: integer
          format: int32
          description: "Year (derived from the object's date, the birth year for Person\
            \ and Sjöman, or the start of the activity period for Juridisk person;\
            \ used for sorting and display)"
          examples:
          - 1920
        topographyId:
          type: integer
          format: int32
          description: Topography ID
          examples:
          - 1
        locationText:
          type: string
          description: Free-text location (the birth parish for Person and Sjöman)
          examples:
          - Sundsvall
        location:
          type: string
          description: Resolved place name from TOPOGRAFI (preferred over locationText
            when set)
          examples:
          - Sundsvall
        nodeId:
          type: integer
          format: int32
          description: "ID of the archive node (arkiv, serie or volym) the object\
            \ was created in, or null when it sits in none. Resolve it with /nodes/{id}\
            \ for the name and the path"
          examples:
          - 19000
        creator:
          $ref: "#/components/schemas/Creator"
    GenderCount:
      type: object
      description: How many objects recording one gender the search matches
      properties:
        gender:
          type: string
          description: "Gender — Man, Kvinna or Okänt, which the gender filter accepts"
          examples:
          - Man
        count:
          type: integer
          format: int64
          description: "Number of matching objects recording that gender, across every\
            \ page. The counters cover every matched row that records a gender at\
            \ all, so they sum to fewer than totalRecords whenever the result also\
            \ holds types that record none."
          examples:
          - 12
    ObjectTypeCount:
      type: object
      description: How many objects of one type the search matches
      properties:
        objectType:
          type: string
          description: "Object type (Foto, Föremål, Film, Ljud, Text, Publikation,\
            \ Person, Juridisk person, Sjöman or Mantal) — the same value each object\
            \ reports and the objectType filter accepts"
          examples:
          - Foto
        count:
          type: integer
          format: int64
          description: "Number of matching objects of that type, across every page"
          examples:
          - 12
    PagedCombinedObjectResponse:
      type: object
      description: Paged combined object response
      properties:
        objects:
          type: array
          items:
            $ref: "#/components/schemas/CombinedObject"
            readOnly: true
        typeCounts:
          type: array
          description: "Total number of matching objects per type (for chip counters),\
            \ ordered by object type. The counts are independent of the current page\
            \ and of the objectType filter — every other filter applies, so a chip\
            \ keeps saying how many objects selecting that type would return."
          items:
            $ref: "#/components/schemas/ObjectTypeCount"
            readOnly: true
        genderCounts:
          type: array
          description: "Total number of matching objects per gender (for chip counters),\
            \ ordered by gender, over the rows that record one. Gender is its own\
            \ dimension — a row is both Person and man — so these overlap with typeCounts\
            \ rather than summing with them. The counts are independent of the current\
            \ page and of the gender filter — every other filter applies, so a chip\
            \ keeps saying how many objects selecting that gender would return."
          items:
            $ref: "#/components/schemas/GenderCount"
            readOnly: true
        categoryCounts:
          type: array
          description: "Total number of matching objects per originator category (for\
            \ chip counters), ordered by category name, over the rows whose originator\
            \ is a categorised legal entity. A dimension of its own like genderCounts,\
            \ so these overlap with typeCounts rather than summing with them, and\
            \ the register rows and objects without an originator are not counted.\
            \ The counts are independent of the current page and of the categoryId\
            \ filter — every other filter applies, so a chip keeps saying how many\
            \ objects selecting that category would return."
          items:
            $ref: "#/components/schemas/CategoryCount"
            readOnly: true
        topographyCounts:
          type: array
          description: "How many matching objects are placed in each topography (place),\
            \ by id and by the name /topographies lists it under, sorted by name —\
            \ for a place picker that can hide the empty places and show the numbers.\
            \ Covers the rows placed in a topography: the object types and the legal\
            \ entities, not the person registers, which hold a parish as free text.\
            \ A dimension of its own like the other counters, so these overlap with\
            \ typeCounts rather than summing with them. The counts are independent\
            \ of the current page and of the topographyId filter — every other filter\
            \ applies, the substring location filter included, so a chip keeps saying\
            \ how many objects selecting that place would return."
          items:
            $ref: "#/components/schemas/TopographyCount"
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
    TopographyCount:
      type: object
      description: How many objects placed in one topography the search matches
      properties:
        topographyId:
          type: integer
          format: int32
          description: "Topography ID, which the topographyId filter accepts"
          examples:
          - 4
        name:
          type: string
          description: "The place's display name, the same one /topographies lists\
            \ it under and an object reports as its location"
          examples:
          - "Kvissleby, Njurunda"
        count:
          type: integer
          format: int64
          description: "Number of matching objects placed in that topography, across\
            \ every page. The counters cover every matched row placed in a topography,\
            \ so they sum to fewer than totalRecords whenever the result also holds\
            \ rows without one."
          examples:
          - 12
    NodeParameters:
      type: object
      description: "Archive and collection node search parameters. All filters are\
        \ optional and combined with AND, except that several values of institutionId,\
        \ categoryId and topographyId are alternatives. Sort on one of: name, startYear,\
        \ stopYear, sortOrder or location."
      properties:
        page:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - 1
          minimum: 1
        limit:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - 15
          minimum: 1
        sortBy:
          type: array
          items:
            type: string
            description: "Property to sort on. location orders by the place a node\
              \ is placed in, nodes without one last whichever way the list runs"
            enum:
            - name
            - startYear
            - stopYear
            - sortOrder
            - location
            examples:
            - name
        sortDirection:
          $ref: "#/components/schemas/Direction"
          description: The sort order direction
          examples:
          - ASC
        query:
          type: string
          description: "Free text search (substring, case-insensitive) across name,\
            \ description and the name of the arkivbildare"
          examples:
          - stadsfullmäktige
        nodeTypeId:
          type: integer
          format: int32
          description: Node type ID
          examples:
          - 2
        nodeType:
          type: string
          description: "Node type by name, matched case-insensitively — Arkiv for\
            \ the archives and collections themselves, which is what a list of arkiv\
            \ och samlingar shows, Serie or Volym for the levels under them"
          examples:
          - Arkiv
        institutionId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the institution (arkivinstitution) holding the archive,\
              \ as listed by /institutions. Repeat the parameter, or comma-separate\
              \ the values, to select several — they are alternatives."
            examples:
            - 3
        categoryId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the arkivbildare's category (verksamhetskategori),\
              \ as listed by /categories. Repeat the parameter, or comma-separate\
              \ the values, to select several — they are alternatives. The sentinel\
              \ every legal entity defaults to is not a category, and naming it alone\
              \ matches nothing."
            examples:
            - 5
        topographyId:
          type: array
          items:
            type: integer
            format: int32
            description: "ID of the place (topografi) the archive is about or comes\
              \ from, as listed by /topographies — the exact counterpart of location,\
              \ for a place picked from that list rather than typed. Repeat the parameter,\
              \ or comma-separate the values, to select several — they are alternatives."
            examples:
            - 4
        location:
          type: string
          description: "Place (substring, case-insensitive; the resolved place name\
            \ or the free-text place)"
          examples:
          - Njurunda
        yearFrom:
          type: integer
          format: int32
          description: Keep nodes whose period reaches this year or later (inclusive)
          examples:
          - 1862
        yearTo:
          type: integer
          format: int32
          description: Keep nodes whose period starts this year or earlier (inclusive)
          examples:
          - 1951
...
    PagedNodeResponse:
      type: object
      description: Paged archive and collection node response
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
            readOnly: true
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
    NodeDetail:
      type: object
      description: Archive or collection node with the path from the root down to
        it
      properties:
        node:
          $ref: "#/components/schemas/Node"
          readOnly: true
        path:
          type: array
          description: "Ancestors of the node, root first, excluding the node itself.\
            \ Empty for a root node."
          items:
            $ref: "#/components/schemas/Node"
            readOnly: true
...
  securitySchemes: {}