# src/main/java/se/sundsvall/memories/api/CombinedObjectResource.java
package se.sundsvall.memories.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import se.sundsvall.dept44.common.validators.annotation.ValidMunicipalityId;
import se.sundsvall.dept44.problem.Problem;
import se.sundsvall.dept44.problem.violations.ConstraintViolationProblem;
import se.sundsvall.memories.api.model.CombinedObjectParameters;
import se.sundsvall.memories.api.model.PagedCombinedObjectResponse;
import se.sundsvall.memories.service.CombinedObjectService;

import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
import static org.springframework.http.MediaType.APPLICATION_PROBLEM_JSON_VALUE;
import static org.springframework.http.ResponseEntity.ok;

@RestController
@Validated
@RequestMapping("/{municipalityId}/objects")
@Tag(name = "Object", description = "Combined object search across all object and register types")
@ApiResponses(value = {
	@ApiResponse(responseCode = "400", description = "Bad request", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(oneOf = {
		Problem.class, ConstraintViolationProblem.class
	}))),
	@ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = APPLICATION_PROBLEM_JSON_VALUE, schema = @Schema(implementation = Problem.class)))
})
class CombinedObjectResource {

	private final CombinedObjectService combinedObjectService;

	CombinedObjectResource(final CombinedObjectService combinedObjectService) {
		this.combinedObjectService = combinedObjectService;
	}

	@GetMapping(produces = APPLICATION_JSON_VALUE)
	@Operation(summary = "Search objects",
		description = "Search across all object types (Foto, Föremål, Film, Ljud, Text, Publikation) and registers (Person, Juridisk person, Sjöman, Mantal) in one call, with global sorting, pagination and counts per type, gender, originator category and place. Pass objectType, gender, categoryId or topographyId to narrow the search; each counter leaves out only its own selection, so the chips stay selectable. Pass nodeId to list what an archive node holds.")
	@ApiResponse(responseCode = "200", description = "Successful operation")
	ResponseEntity<PagedCombinedObjectResponse> searchObjects(
		@PathVariable @ValidMunicipalityId final String municipalityId,
		@Valid final CombinedObjectParameters parameters) {

		return ok(combinedObjectService.search(parameters));
	}
}